change return type

+++ src/cn/trinea/android/common/util/ScreenUtils.java
@@ -32,11 +32,11 @@ public class ScreenUtils {
         return px / context.getResources().getDisplayMetrics().density;
     }
 
-    public static float dpToPxInt(Context context, float dp) {
+    public static int dpToPxInt(Context context, float dp) {
         return (int)(dpToPx(context, dp) + 0.5f);
     }
 
-    public static float pxToDpCeilInt(Context context, float px) {
+    public static int pxToDpCeilInt(Context context, float px) {
         return (int)(pxToDp(context, px) + 0.5f);
     }
 }
